Create to count the number of words in a string using HashMap.
split method is used on the input string str with the regular expression "\\s+". This regular expression splits the string into an array of words based on one or more whitespace characters ,
The length of the resulting array of words is returned, which represents the number of words in the original string.